+++ src/DocumentationParser.cpp
@@ -131,7 +131,7 @@ std::optional<std::string> printDocumentation(const Luau::DocumentationDatabase&
             if (!basic->learnMoreLink.empty())
                 result += "\n\n[Learn More](" + basic->learnMoreLink + ")";
             if (!basic->codeSample.empty())
-                result += "\n\n" + codeBlock("lua", basic->codeSample);
+                result += "\n\n" + codeBlock("luau", basic->codeSample);
         }
         else if (auto* func = documentation->get_if<Luau::FunctionDocumentation>())
         {
@@ -139,7 +139,7 @@ std::optional<std::string> printDocumentation(const Luau::DocumentationDatabase&
             if (!func->learnMoreLink.empty())
                 result += "\n\n[Learn More](" + func->learnMoreLink + ")";
             if (!func->codeSample.empty())
-                result += "\n\n" + codeBlock("lua", func->codeSample);
+                result += "\n\n" + codeBlock("luau", func->codeSample);
         }
         else if (auto* overloaded = documentation->get_if<Luau::OverloadedFunctionDocumentation>())
         {
@@ -159,7 +159,7 @@ std::optional<std::string> printDocumentation(const Luau::DocumentationDatabase&
             if (!tbl->learnMoreLink.empty())
                 result += "\n\n[Learn More](" + tbl->learnMoreLink + ")";
             if (!tbl->codeSample.empty())
-                result += "\n\n" + codeBlock("lua", tbl->codeSample);
+                result += "\n\n" + codeBlock("luau", tbl->codeSample);
         }
         return result;
     }