+++ CHANGELOG.md
@@ -10,6 +10,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Sync to upstream Luau 0.597
 - Prioritise `game:GetService()` as the first autocompletion entry when typing `game:`
+- Code blocks in hover and documentation now use `luau` as the syntax highlighting
 
 ### Fixed